backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
deepface==0.0.99
exceptiongroup==1.3.1
fastapi==0.128.6
filelock==3.25.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h11==0.16.0
h5py==3.16.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
llvmlite==0.46.0
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numba==0.63.1
numpy==2.1.3
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.5.1
pooch==1.9.0
protobuf==3.20.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMatting==1.1.15
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
retina-face==0.0.17
rich==14.3.3
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tf_keras==2.19.0
tifffile==2025.5.10
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.6
wrapt==2.1.2
cloudinary==1.38.0
pymongo[srv]==4.6.1
